refactoring files organization
