+++ carla_gym/core/constants.py
@@ -83,27 +83,6 @@ RETRIES_ON_ERROR = 2
 # Dummy Z coordinate to use when we only care about (x, y)
 GROUND_Z = 22
 
-DISCRETE_ACTIONS = {
-    # coast
-    0: [0.0, 0.0],
-    # turn left
-    1: [0.0, -0.5],
-    # turn right
-    2: [0.0, 0.5],
-    # forward
-    3: [1.0, 0.0],
-    # brake
-    4: [-0.5, 0.0],
-    # forward left
-    5: [0.5, -0.05],
-    # forward right
-    6: [0.5, 0.05],
-    # brake left
-    7: [-0.5, -0.5],
-    # brake right
-    8: [-0.5, 0.5],
-}
-
 WEATHERS = {
     "Default": carla.WeatherParameters.Default,
     "ClearNoon": carla.WeatherParameters.ClearNoon,
