+++ carla_gym/core/controllers/camera_manager.py
