+++ carla_gym/core/controllers/pedestrian_manager.py
